+++ tests/test_dataframe.py
@@ -1,7 +1,6 @@
 import pytest
 from mhctools import NetMHC
 from topiary import TopiaryPredictor
-from .data import cancer_test_variants
 
 alleles = [
     "A02:01",
@@ -17,7 +16,7 @@ def mhc_model():
     return NetMHC(alleles=alleles, default_peptide_lengths=[8, 9, 10])
 
 
-def test_epitopes_to_dataframe_transcript_expression(mhc_model):
+def test_epitopes_to_dataframe_transcript_expression(mhc_model, cancer_test_variants):
     predictor = TopiaryPredictor(mhc_model=mhc_model, only_novel_epitopes=False)
     df = predictor.predict_from_variants(
         variants=cancer_test_variants,
@@ -36,7 +35,7 @@ def test_epitopes_to_dataframe_transcript_expression(mhc_model):
     ).all(), "Invalid FPKM values in DataFrame transcript_expression column"
 
 
-def test_epitopes_to_dataframe_gene_expression(mhc_model):
+def test_epitopes_to_dataframe_gene_expression(mhc_model, cancer_test_variants):
     predictor = TopiaryPredictor(mhc_model=mhc_model, only_novel_epitopes=False)
 
     df = predictor.predict_from_variants(