+++ tests/test_effect_expression_filters.py
@@ -1,26 +1,9 @@
-from .data import (
-    cancer_test_variants,
-    cancer_test_variant_gene_ids,
-    cancer_test_variant_transcript_ids,
-)
 from topiary.filters import apply_effect_expression_filters
-
-cancer_test_effects = cancer_test_variants.effects()
 
 DEFAULT_FPKM = 1.0
 
-# associate every gene ID with 1.0 FPKM
-gene_expression_dict = {
-    gene_id: DEFAULT_FPKM for gene_id in cancer_test_variant_gene_ids
-}
 
-# associate every transcript with 1.0 FPKM
-transcript_expression_dict = {
-    transcript_id: DEFAULT_FPKM for transcript_id in cancer_test_variant_transcript_ids
-}
-
-
-def test_apply_effect_gene_expression_below_threshold():
+def test_apply_effect_gene_expression_below_threshold(cancer_test_effects, gene_expression_dict):
     filtered = apply_effect_expression_filters(
         cancer_test_effects,
         gene_expression_dict=gene_expression_dict,
@@ -33,7 +16,7 @@ def test_apply_effect_gene_expression_below_threshold():
     ), "All variants should have been filtered out but got: %s" % (filtered,)
 
 
-def test_apply_effect_gene_expression_above_threshold():
+def test_apply_effect_gene_expression_above_threshold(cancer_test_effects, gene_expression_dict):
     filtered = apply_effect_expression_filters(
         cancer_test_effects,
         gene_expression_dict=gene_expression_dict,
@@ -46,7 +29,7 @@ def test_apply_effect_gene_expression_above_threshold():
     ), "Expected %s effects but got %s" % (len(cancer_test_effects), len(filtered))
 
 
-def test_apply_effect_gene_expression_equal_threshold():
+def test_apply_effect_gene_expression_equal_threshold(cancer_test_effects, gene_expression_dict):
     # expect genes with expression at threshold to NOT get filtered
     filtered = apply_effect_expression_filters(
         cancer_test_effects,
@@ -60,7 +43,7 @@ def test_apply_effect_gene_expression_equal_threshold():
     ), "Expected %s effects but got %s" % (len(cancer_test_effects), len(filtered))
 
 
-def test_apply_effect_transcript_expression_below_threshold():
+def test_apply_effect_transcript_expression_below_threshold(cancer_test_effects, transcript_expression_dict):
     filtered = apply_effect_expression_filters(
         cancer_test_effects,
         gene_expression_dict=None,
@@ -73,7 +56,7 @@ def test_apply_effect_transcript_expression_below_threshold():
     ), "All effects should have been filtered out but got: %s" % (filtered,)
 
 
-def test_apply_effect_transcript_expression_above_threshold():
+def test_apply_effect_transcript_expression_above_threshold(cancer_test_effects, transcript_expression_dict):
     filtered = apply_effect_expression_filters(
         cancer_test_effects,
         gene_expression_dict=None,
@@ -86,7 +69,7 @@ def test_apply_effect_transcript_expression_above_threshold():
     ), "Expected %s effects but got %s" % (len(cancer_test_effects), len(filtered))
 
 
-def test_apply_effect_transcript_expression_equal_threshold():
+def test_apply_effect_transcript_expression_equal_threshold(cancer_test_effects, transcript_expression_dict):
     # expect transcripts with expression at threshold to NOT be filtered
     filtered = apply_effect_expression_filters(
         cancer_test_effects,